# Phase 1: Observer Pattern Architecture & Event-Driven Communication